+++ src/rogallo/data/location_history.py
@@ -13,11 +13,16 @@ from typing import Self
 from bagofstuff.history import RecencyHistory
 
 ##############################################################################
+# Port79 imports.
+from port79 import FingerURI
+
+##############################################################################
 # Wasat imports.
 from wasat import GeminiURI
 
 ##############################################################################
 # Local imports.
+from ..preflight import is_finger_uri, is_gemini_uri
 from ..types import GeminiLocation
 from .locations import data_dir
 
@@ -46,7 +51,9 @@ class LocationVisit:
             The visit as a JSON-serialisable dictionary.
         """
         return {
-            "type": "uri" if isinstance(self.location, GeminiURI) else "path",
+            "type": "uri"
+            if isinstance(self.location, (FingerURI, GeminiURI))
+            else "path",
             "location": str(self.location),
             "timestamp": self.timestamp.isoformat(),
         }
@@ -61,8 +68,14 @@ class LocationVisit:
         Returns:
             The visit.
         """
+        maker: type[FingerURI | GeminiURI | Path] = Path
+        if data["type"] == "uri":
+            if is_gemini_uri(data["location"]):
+                maker = GeminiURI
+            elif is_finger_uri(data["location"]):
+                maker = FingerURI
         return cls(
-            (GeminiURI if data["type"] == "uri" else Path)(data["location"]),
+            maker(data["location"]),
             datetime.fromisoformat(data["timestamp"]),
         )